Refactor teleport event publishers
