+++ internal/starknet/sequencer.go
@@ -52,10 +52,13 @@ func (s *Sequencer) getBlock(ctx context.Context, blockNumber string) (*Block, e
 	}
 	req = req.WithContext(ctx)
 	res, err := s.httpClient.Do(req)
+	defer res.Body.Close()
 	if err != nil {
 		return nil, err
 	}
-	defer res.Body.Close()
+	if res.StatusCode != http.StatusOK {
+		return nil, fmt.Errorf("unexpected status code: %d", res.StatusCode)
+	}
 	body, err := io.ReadAll(res.Body)
 	if err != nil {
 		return nil, err
