--- pkg/event/publisher/ethereum/listener_test.go
@@ -1,84 +1,0 @@
-//  Copyright (C) 2020 Maker Ecosystem Growth Holdings, INC.
-//
-//  This program is free software: you can redistribute it and/or modify
-//  it under the terms of the GNU Affero General Public License as
-//  published by the Free Software Foundation, either version 3 of the
-//  License, or (at your option) any later version.
-//
-//  This program is distributed in the hope that it will be useful,
-//  but WITHOUT ANY WARRANTY; without even the implied warranty of
-//  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-//  GNU Affero General Public License for more details.
-//
-//  You should have received a copy of the GNU Affero General Public License
-//  along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
-package ethereum
-
-import (
-	"context"
-	"testing"
-	"time"
-
-	geth "github.com/ethereum/go-ethereum"
-	"github.com/ethereum/go-ethereum/common"
-	"github.com/ethereum/go-ethereum/core/types"
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/mock"
-
-	"github.com/chronicleprotocol/oracle-suite/pkg/ethereum"
-	"github.com/chronicleprotocol/oracle-suite/pkg/ethereum/geth/mocks"
-	"github.com/chronicleprotocol/oracle-suite/pkg/log/null"
-)
-
-var listenerTestAddress = common.HexToAddress("0x2d800d93b065ce011af83f316cef9f0d005b0aa4")
-var listenerTestTopic1 = ethereum.HexToHash("0x0162851814eed5360ac17d1b3d942e6619fa2d803de71e7159ed9bebf724072a")
-var listenerTestTopic2 = ethereum.HexToHash("0x618b439c5646a69d0ee5bf11275cce6932a3eaca08b9ad1c68b4531001560de7")
-
-func Test_logListener(t *testing.T) {
-	ctx, cancelFunc := context.WithCancel(context.Background())
-	defer cancelFunc()
-	cli := &mocks.EthClient{}
-	lis := &logListener{
-		client:      cli,
-		addresses:   []common.Address{listenerTestAddress},
-		topics:      [][]common.Hash{{listenerTestTopic1}, {listenerTestTopic2}},
-		interval:    time.Millisecond * 100,
-		blocksDelta: []uint64{10},
-		blocksLimit: 15,
-		logCh:       make(chan types.Log),
-		logger:      null.New(),
-	}
-
-	// During the first call we are expecting to fetch up to blocksLimit.
-	cli.On("BlockNumber", ctx).Return(uint64(42), nil).Once()
-	cli.On("FilterLogs", ctx, mock.Anything).Return([]types.Log{{Index: 1}, {Index: 2}}, nil).Once().Run(func(args mock.Arguments) {
-		fq := args.Get(1).(geth.FilterQuery)
-		assert.Equal(t, uint64(18), fq.FromBlock.Uint64())
-		assert.Equal(t, uint64(32), fq.ToBlock.Uint64())
-		assert.Equal(t, []common.Address{listenerTestAddress}, fq.Addresses)
-		assert.Equal(t, [][]common.Hash{{listenerTestTopic1}, {listenerTestTopic2}}, fq.Topics)
-	})
-	// During the second call, we expect to fetch blocks between the last
-	// fetched one and the current one minus the value of blocksDelta..
-	cli.On("BlockNumber", ctx).Return(uint64(52), nil).Once()
-	cli.On("FilterLogs", ctx, mock.Anything).Return([]types.Log{{Index: 3}, {Index: 4}}, nil).Once().Run(func(args mock.Arguments) {
-		fq := args.Get(1).(geth.FilterQuery)
-		assert.Equal(t, uint64(33), fq.FromBlock.Uint64())
-		assert.Equal(t, uint64(42), fq.ToBlock.Uint64())
-		assert.Equal(t, []common.Address{listenerTestAddress}, fq.Addresses)
-		assert.Equal(t, [][]common.Hash{{listenerTestTopic1}, {listenerTestTopic2}}, fq.Topics)
-	})
-	// If the latest block number is the same as the one during previous call,
-	// the fetchLogs method should do nothing.
-	cli.On("BlockNumber", ctx).Return(uint64(52), nil).Once()
-
-	// Start listener and collect logs:
-	lis.start(ctx)
-	var logs []types.Log
-	for len(logs) < 4 {
-		logs = append(logs, <-lis.logCh)
-		time.Sleep(time.Millisecond * 10)
-	}
-	assert.Len(t, logs, 4)
-}
